Backport concurrent modification bug fix (PRESS-20110)
In our Java 11 upgrade we need to use this version of the zinc compiler.
It's the last version compiled with Scala 2.13.3. Unfortunately, it
includes this [bug][1]. The was fixed [here][2], and we are just
backporting the fix for now. Soon (TM) we hope to get back to the zinc's
main branch once we upgrade our Scala version.

Backporting also means we have to compile the package from source and
publish it to our internal Maven repository. You can do that with the
following steps:

1. Clone the repository and checkout this commit
2. Run the `publishM2` SBT task. This will compile and store the jars in `~/.m2`.
3. Zip up the repository with the following command: `cd ~/.m2/repository && zip -r repo.zip org`.
4. Upload this zip to the `ext-release-local` repository using the JFrog
   web app here: https://repo.lucidpressutil.com/ui/repos/tree/General/ext-release-local
   a. You'll want to log in as an admin
   b. There is a repoistory import option here: https://repo.lucidpressutil.com/ui/admin/artifactory/import_export/repositories

[1]: https://github.com/sbt/zinc/issues/974
[2]: https://github.com/sbt/zinc/pull/975